+++ pkg/common.go
@@ -1,0 +1,6 @@
+package pkg
+
+const (
+	// Name of the lib. Used as a prefix for kt_errors.Fault sources and Logging
+	LIB_NAME = "keytiles.lib.utils"
+)