+++ tests/string_varextract_test.go
@@ -3,8 +3,8 @@ package kt_utils_test
 import (
 	"testing"
 
-	"github.com/keytiles/lib-sets-golang/ktsets"
-	"github.com/keytiles/lib-utils-golang/pkg/kt_utils"
+	"github.com/keytiles/lib-sets-golang/v2/pkg/kt_sets"
+	"github.com/keytiles/lib-utils-golang/v2/pkg/kt_utils"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -14,5 +14,5 @@ func TestExtractVariableNamesFromString(t *testing.T) {
 	// ---- WHEN
 	vars := kt_utils.StringExtractVariableNames(str)
 	// ---- THEN
-	assert.True(t, vars.Equals(ktsets.NewSet("var1", "var2")))
+	assert.True(t, vars.Equals(kt_sets.NewSet("var1", "var2")))
 }